# Performance: short-circuit Bind0 for collections